Take pg version out of commission. Investigate generators in conjunction with psycopg cursors to arrive at more elegant query syntax
